Delete ClusterRoleBinding when Mirror Maker 2 is deleted
